Add two new fuzzers for Exact minimum and maximum values in Numberic fields
This involved a bit more refactoring in order to pass the fuzzedField within the hasBoudartyDefined() methods.
When a field is sent through refernce data, there is no point in fuzzing it with a min or max value, as
the test will fail with a 4XX, while CATS expects to get a 2XX as it sends a structural valid value.